Move functions to new module
* I'll probably organize them somewhere else in the future,
but for now I need them out of ui.py before splitting it.
